Initialized Navbar
Only for registration only, as auth has not been configured. Will add temp drawers next.
